+++ src/attractors/lotka_volterra.py
@@ -56,7 +56,7 @@ OR CORRECTION.
 from typing import Tuple
 
 
-def lotka_volterra(x: int = 0, y: int = 0, z: int = 1) -> Tuple[int, int, int]:
+def lotka_volterra(x: float = 0, y: float = 0, z: float = 1) -> Tuple[float, float, float]:
     """Calculate the next coordinate X, Y, Z for Lotka–Volterra
 
     Parameters
