+++ src/attractors/wang.py
@@ -46,7 +46,7 @@ OR CORRECTION.
 from typing import Tuple
 
 
-def wang(x: int = 0, y: int = 0, z: int = 1) -> Tuple[int, int, int]:
+def wang(x: float = 0, y: float = 0, z: float = 1) -> Tuple[float, float, float]:
     """Calculate the next coordinate X, Y, Z for 3rd-order Wang Attractor
 
     Parameters
